added swagger config

+++ src/main/java/com/reactive/examples/dto/UserDepartmentDTO.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @Builder
@@ -13,7 +15,7 @@ public class UserDepartmentDTO {
     private Integer userId;
     private String userName;
     private int age;
-    private double salary;
+    private BigDecimal salary;
     private Integer departmentId;
     private String departmentName;
     private String loc;
